Move validateNodeAddress to DisputeValidation
