+++ core/src/main/java/bisq/core/support/dispute/DisputeManager.java
@@ -273,8 +273,8 @@ public abstract class DisputeManager<T extends DisputeList<Dispute>> extends Sup
         disputes.forEach(dispute -> {
             try {
                 TradeDataValidation.validateDonationAddress(dispute, dispute.getDonationAddressOfDelayedPayoutTx(), daoFacade);
-                TradeDataValidation.validateNodeAddress(dispute, dispute.getContract().getBuyerNodeAddress(), config);
-                TradeDataValidation.validateNodeAddress(dispute, dispute.getContract().getSellerNodeAddress(), config);
+                DisputeValidation.validateNodeAddress(dispute, dispute.getContract().getBuyerNodeAddress(), config);
+                DisputeValidation.validateNodeAddress(dispute, dispute.getContract().getSellerNodeAddress(), config);
             } catch (DisputeValidation.AddressException | DisputeValidation.NodeAddressException e) {
                 log.error(e.toString());
                 validationExceptions.add(e);
@@ -370,8 +370,8 @@ public abstract class DisputeManager<T extends DisputeList<Dispute>> extends Sup
         try {
             TradeDataValidation.validateDonationAddress(dispute.getDonationAddressOfDelayedPayoutTx(), daoFacade);
             TradeDataValidation.testIfDisputeTriesReplay(dispute, disputeList.getList());
-            TradeDataValidation.validateNodeAddress(dispute, dispute.getContract().getBuyerNodeAddress(), config);
-            TradeDataValidation.validateNodeAddress(dispute, dispute.getContract().getSellerNodeAddress(), config);
+            DisputeValidation.validateNodeAddress(dispute, dispute.getContract().getBuyerNodeAddress(), config);
+            DisputeValidation.validateNodeAddress(dispute, dispute.getContract().getSellerNodeAddress(), config);
         } catch (DisputeValidation.AddressException |
                  DisputeValidation.DisputeReplayException |
                  DisputeValidation.NodeAddressException e) {
